refactor: rename pkg/mat to pkg/model
Package named after domain concept (model parameters/activations),
not data structure (mat). Updates imports, Makefile, and tests.

model.New() reads as 'new model matrix' — the package carries context.
