+++ coverage_mat.out
@@ -1,0 +1,46 @@
+mode: set
+dipirona/pkg/mat/matrix.go:17.37,20.14 3 1
+dipirona/pkg/mat/matrix.go:20.14,22.3 1 1
+dipirona/pkg/mat/matrix.go:24.2,25.24 2 1
+dipirona/pkg/mat/matrix.go:25.24,28.3 2 1
+dipirona/pkg/mat/matrix.go:30.2,34.3 1 1
+dipirona/pkg/mat/matrix.go:38.42,40.2 1 1
+dipirona/pkg/mat/matrix.go:43.36,45.22 2 1
+dipirona/pkg/mat/matrix.go:45.22,47.3 1 1
+dipirona/pkg/mat/matrix.go:49.2,49.30 1 1
+dipirona/pkg/mat/matrix.go:49.30,50.31 1 1
+dipirona/pkg/mat/matrix.go:50.31,52.4 1 1
+dipirona/pkg/mat/matrix.go:55.2,59.3 1 1
+dipirona/pkg/mat/matrix.go:64.47,65.26 1 1
+dipirona/pkg/mat/matrix.go:65.26,66.115 1 1
+dipirona/pkg/mat/matrix.go:69.2,70.22 2 1
+dipirona/pkg/mat/matrix.go:70.22,72.3 1 1
+dipirona/pkg/mat/matrix.go:74.2,74.30 1 1
+dipirona/pkg/mat/matrix.go:74.30,75.35 1 1
+dipirona/pkg/mat/matrix.go:75.35,77.32 2 1
+dipirona/pkg/mat/matrix.go:77.32,79.5 1 1
+dipirona/pkg/mat/matrix.go:80.4,80.20 1 1
+dipirona/pkg/mat/matrix.go:84.2,88.3 1 1
+dipirona/pkg/mat/matrix.go:94.77,95.50 1 1
+dipirona/pkg/mat/matrix.go:95.50,96.112 1 1
+dipirona/pkg/mat/matrix.go:99.2,100.22 2 1
+dipirona/pkg/mat/matrix.go:100.22,102.3 1 1
+dipirona/pkg/mat/matrix.go:104.2,104.30 1 1
+dipirona/pkg/mat/matrix.go:104.30,105.31 1 1
+dipirona/pkg/mat/matrix.go:105.31,107.4 1 1
+dipirona/pkg/mat/matrix.go:110.2,114.3 1 1
+dipirona/pkg/mat/matrix.go:119.54,121.22 2 1
+dipirona/pkg/mat/matrix.go:121.22,123.3 1 1
+dipirona/pkg/mat/matrix.go:125.2,125.30 1 1
+dipirona/pkg/mat/matrix.go:125.30,126.31 1 1
+dipirona/pkg/mat/matrix.go:126.31,128.4 1 1
+dipirona/pkg/mat/matrix.go:131.2,135.3 1 1
+dipirona/pkg/mat/matrix.go:139.33,142.30 3 1
+dipirona/pkg/mat/matrix.go:142.30,143.12 1 1
+dipirona/pkg/mat/matrix.go:143.12,145.4 1 1
+dipirona/pkg/mat/matrix.go:146.3,147.31 2 1
+dipirona/pkg/mat/matrix.go:147.31,148.13 1 1
+dipirona/pkg/mat/matrix.go:148.13,150.5 1 1
+dipirona/pkg/mat/matrix.go:151.4,151.50 1 1
+dipirona/pkg/mat/matrix.go:153.3,153.21 1 1
+dipirona/pkg/mat/matrix.go:155.2,156.19 2 1
